projeto_asma/requirements.txt
Django==3.0.6
django-cors-headers==3.2.1
djangorestframework==3.11.0
django-rest-knox